Skip slow calculation of reference C6 (#39)
The loops in the `_load_c6` function are actually really slow and create
a bottleneck for the whole calculation. Since the output of this
function is not dependent on the system, we only use it to store the
tensor (".pt" file) and now skip the calculation and directly load the
tensor from the "reference-c6.pt" file.

This gives a speed-up of more than one order of magnitude for small
systems.